requirements.txt
numpy
sympy
PyYAML
datasets
wandb
latex2sympy2
trl
unsloth
tqdm

